requirements.txt
wheel==0.23.0
python-logstash=0.4.2
